chore: move localcatalog package to datacatalogprovider
