+++ cli/internal/providers/datacatalog/state/customtype.go
@@ -1,7 +1,7 @@
 package state
 
 import (
-	"github.com/rudderlabs/rudder-iac/cli/internal/localcatalog"
+	"github.com/rudderlabs/rudder-iac/cli/internal/providers/datacatalog/localcatalog"
 	"github.com/rudderlabs/rudder-iac/cli/internal/syncer/resources"
 )
