+++ cli/internal/providers/datacatalog/state/customtype_test.go
@@ -3,7 +3,7 @@ package state
 import (
 	"testing"
 
-	"github.com/rudderlabs/rudder-iac/cli/internal/localcatalog"
+	"github.com/rudderlabs/rudder-iac/cli/internal/providers/datacatalog/localcatalog"
 	"github.com/rudderlabs/rudder-iac/cli/internal/syncer/resources"
 	"github.com/stretchr/testify/assert"
 )
